frontend/requirements.txt
streamlit==1.49.1
requests==2.32.5
python-dotenv==1.1.1
